chore: update code structure (#1139)
* chore: update code structure

* chore: update
