+++ Dockerfile
@@ -15,7 +15,7 @@ RUN apk update && apk add --no-cache gcc musl-dev
 COPY . .
 COPY --from=frontend /frontend-build/dist ./server/dist
 
-RUN go build -o memos ./bin/server/main.go
+RUN go build -o memos ./main.go
 
 # Make workspace with above generated files.
 FROM alpine:3.16 AS monolithic
