+++ scripts/build.sh
@@ -8,6 +8,6 @@ cd "$(dirname "$0")/../"
 
 echo "Start building backend..."
 
-go build -o ./build/memos ./bin/server/main.go
+go build -o ./build/memos ./main.go
 
 echo "Backend built!"
